moved vehicle simulation to simple sensor simulator
