+++ simulation/simple_sensor_simulator/include/simple_sensor_simulator/vehicle_simulation/vehicle_model/sim_model_delay_steer_vel.hpp
@@ -12,17 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef SIMPLE_PLANNING_SIMULATOR__VEHICLE_MODEL__SIM_MODEL_DELAY_STEER_ACC_GEARED_HPP_
-#define SIMPLE_PLANNING_SIMULATOR__VEHICLE_MODEL__SIM_MODEL_DELAY_STEER_ACC_GEARED_HPP_
+#ifndef SIMPLE_PLANNING_SIMULATOR__VEHICLE_MODEL__SIM_MODEL_DELAY_STEER_VEL_HPP_
+#define SIMPLE_PLANNING_SIMULATOR__VEHICLE_MODEL__SIM_MODEL_DELAY_STEER_VEL_HPP_
 
 #include <deque>
 #include <eigen3/Eigen/Core>
 #include <eigen3/Eigen/LU>
 #include <iostream>
 #include <queue>
-#include <traffic_simulator/vehicle_simulation/vehicle_model/sim_model_interface.hpp>
-
-class SimModelDelaySteerAccGeared : public SimModelInterface
+#include <simple_sensor_simulator/vehicle_simulation/vehicle_model/sim_model_interface.hpp>
+/**
+ * @class SimModelDelaySteerVel
+ * @brief calculate delay steering dynamics
+ */
+class SimModelDelaySteerVel : public SimModelInterface
 {
 public:
   /**
@@ -33,20 +36,20 @@ public:
    * @param [in] steer_rate_lim steering angular velocity limit [rad/ss]
    * @param [in] wheelbase vehicle wheelbase length [m]
    * @param [in] dt delta time information to set input buffer for delay
-   * @param [in] acc_delay time delay for accel command [s]
-   * @param [in] acc_time_constant time constant for 1D model of accel dynamics
+   * @param [in] vx_delay time delay for velocity command [s]
+   * @param [in] vx_time_constant time constant for 1D model of velocity dynamics
    * @param [in] steer_delay time delay for steering command [s]
    * @param [in] steer_time_constant time constant for 1D model of steering dynamics
    */
-  SimModelDelaySteerAccGeared(
+  SimModelDelaySteerVel(
     double vx_lim, double steer_lim, double vx_rate_lim, double steer_rate_lim, double wheelbase,
-    double dt, double acc_delay, double acc_time_constant, double steer_delay,
+    double dt, double vx_delay, double vx_time_constant, double steer_delay,
     double steer_time_constant);
 
   /**
-   * @brief default destructor
+   * @brief destructor
    */
-  ~SimModelDelaySteerAccGeared() = default;
+  ~SimModelDelaySteerVel() = default;
 
 private:
   const double MIN_TIME_CONSTANT;  //!< @brief minimum time constant
@@ -57,26 +60,28 @@ private:
     YAW,
     VX,
     STEER,
-    ACCX,
   };
   enum IDX_U {
-    ACCX_DES = 0,
+    VX_DES = 0,
     STEER_DES,
-    DRIVE_SHIFT,
   };
 
-  const double vx_lim_;          //!< @brief velocity limit [m/s]
-  const double vx_rate_lim_;     //!< @brief acceleration limit [m/ss]
+  const double vx_lim_;          //!< @brief velocity limit
+  const double vx_rate_lim_;     //!< @brief acceleration limit
   const double steer_lim_;       //!< @brief steering limit [rad]
   const double steer_rate_lim_;  //!< @brief steering angular velocity limit [rad/s]
   const double wheelbase_;       //!< @brief vehicle wheelbase length [m]
+  double prev_vx_ = 0.0;
+  double current_ax_ = 0.0;
 
-  std::deque<double> acc_input_queue_;    //!< @brief buffer for accel command
-  std::deque<double> steer_input_queue_;  //!< @brief buffer for steering command
-  const double acc_delay_;                //!< @brief time delay for accel command [s]
-  const double acc_time_constant_;        //!< @brief time constant for accel dynamics
-  const double steer_delay_;              //!< @brief time delay for steering command [s]
-  const double steer_time_constant_;      //!< @brief time constant for steering dynamics
+  std::deque<double> vx_input_queue_;     //!< @brief buffer for velocity command
+  std::deque<double> steer_input_queue_;  //!< @brief buffer for angular velocity command
+  const double vx_delay_;                 //!< @brief time delay for velocity command [s]
+  const double vx_time_constant_;
+  //!< @brief time constant for 1D model of velocity dynamics
+  const double steer_delay_;  //!< @brief time delay for angular-velocity command [s]
+  const double
+    steer_time_constant_;  //!< @brief time constant for 1D model of angular-velocity dynamics
 
   /**
    * @brief set queue buffer for input command
@@ -100,7 +105,7 @@ private:
   double getYaw() override;
 
   /**
-   * @brief get vehicle velocity vx
+   * @brief get vehicle longitudinal velocity
    */
   double getVx() override;
 
@@ -131,22 +136,11 @@ private:
   void update(const double & dt) override;
 
   /**
-   * @brief calculate derivative of states with time delay steering model
+   * @brief calculate derivative of states with delay steering model
    * @param [in] state current model state
    * @param [in] input input vector to model
    */
   Eigen::VectorXd calcModel(const Eigen::VectorXd & state, const Eigen::VectorXd & input) override;
-
-  /**
-   * @brief update state considering current gear
-   * @param [in] state current state
-   * @param [in] prev_state previous state
-   * @param [in] gear current gear (defined in autoware_auto_msgs/GearCommand)
-   * @param [in] dt delta time to update state
-   */
-  void updateStateWithGear(
-    Eigen::VectorXd & state, const Eigen::VectorXd & prev_state, const uint8_t gear,
-    const double dt);
 };
 
-#endif  // SIMPLE_PLANNING_SIMULATOR__VEHICLE_MODEL__SIM_MODEL_DELAY_STEER_ACC_GEARED_HPP_
+#endif  // SIMPLE_PLANNING_SIMULATOR__VEHICLE_MODEL__SIM_MODEL_DELAY_STEER_VEL_HPP_
