+++ simulation/simple_sensor_simulator/include/simple_sensor_simulator/vehicle_simulation/vehicle_model/sim_model_ideal_steer_vel.hpp
@@ -18,7 +18,7 @@
 #include <eigen3/Eigen/Core>
 #include <eigen3/Eigen/LU>
 #include <iostream>
-#include <traffic_simulator/vehicle_simulation/vehicle_model/sim_model_interface.hpp>
+#include <simple_sensor_simulator/vehicle_simulation/vehicle_model/sim_model_interface.hpp>
 
 /**
  * @class SimModelIdealSteerVel
