+++ simulation/traffic_simulator/include/traffic_simulator/entity/ego_entity.hpp
@@ -25,7 +25,6 @@
 #include <string>
 #include <traffic_simulator/api/configuration.hpp>
 #include <traffic_simulator/entity/vehicle_entity.hpp>
-#include <traffic_simulator/vehicle_simulation/vehicle_model/sim_model.hpp>
 #include <traffic_simulator_msgs/msg/entity_type.hpp>
 #include <vector>
